@@ -3,13 +3,13 @@
 namespace VladimirYuldashev\Flysystem;
 
 use DateTime;
+use League\Flysystem\Adapter\AbstractFtpAdapter;
+use League\Flysystem\AdapterInterface;
+use League\Flysystem\Config;
+use League\Flysystem\Util;
+use League\Flysystem\Util\MimeType;
 use Normalizer;
 use RuntimeException;
-use League\Flysystem\Util;
-use League\Flysystem\Config;
-use League\Flysystem\Util\MimeType;
-use League\Flysystem\AdapterInterface;
-use League\Flysystem\Adapter\AbstractFtpAdapter;
 
 class CurlFtpAdapter extends AbstractFtpAdapter
 {
@@ -32,6 +32,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
         'proxyUsername',
         'proxyPassword',
         'verbose',
+        'enableTimestampsOnUnixListings'
     ];
 
     /** @var Curl */
@@ -84,7 +85,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * @param bool $ssl
      */
-    public function setSsl($ssl)
+    public function setSsl($ssl): void
     {
         $this->ssl = (bool) $ssl;
     }
@@ -92,7 +93,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * @param int $sslVerifyPeer
      */
-    public function setSslVerifyPeer($sslVerifyPeer)
+    public function setSslVerifyPeer($sslVerifyPeer): void
     {
         $this->sslVerifyPeer = $sslVerifyPeer;
     }
@@ -100,7 +101,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * @param int $sslVerifyHost
      */
-    public function setSslVerifyHost($sslVerifyHost)
+    public function setSslVerifyHost($sslVerifyHost): void
     {
         $this->sslVerifyHost = $sslVerifyHost;
     }
@@ -108,7 +109,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * @param bool $utf8
      */
-    public function setUtf8($utf8)
+    public function setUtf8($utf8): void
     {
         $this->utf8 = (bool) $utf8;
     }
@@ -140,7 +141,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * @param string $proxyHost
      */
-    public function setProxyHost($proxyHost)
+    public function setProxyHost($proxyHost): void
     {
         $this->proxyHost = $proxyHost;
     }
@@ -156,7 +157,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * @param int $proxyPort
      */
-    public function setProxyPort($proxyPort)
+    public function setProxyPort($proxyPort): void
     {
         $this->proxyPort = $proxyPort;
     }
@@ -172,7 +173,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * @param string $proxyUsername
      */
-    public function setProxyUsername($proxyUsername)
+    public function setProxyUsername($proxyUsername): void
     {
         $this->proxyUsername = $proxyUsername;
     }
@@ -188,7 +189,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * @param string $proxyPassword
      */
-    public function setProxyPassword($proxyPassword)
+    public function setProxyPassword($proxyPassword): void
     {
         $this->proxyPassword = $proxyPassword;
     }
@@ -204,7 +205,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * Establish a connection.
      */
-    public function connect()
+    public function connect(): void
     {
         $this->connection = new Curl();
         $this->connection->setOptions([
@@ -253,7 +254,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * Close the connection.
      */
-    public function disconnect()
+    public function disconnect(): void
     {
         if ($this->connection !== null) {
             $this->connection = null;
@@ -268,7 +269,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
      */
     public function isConnected()
     {
-        return $this->connection !== null && !$this->hasConnectionReachedTimeout();
+        return $this->connection !== null && ! $this->hasConnectionReachedTimeout();
     }
 
     /**
@@ -375,13 +376,13 @@ class CurlFtpAdapter extends AbstractFtpAdapter
         $connection = $this->getConnection();
 
         $response = $this->rawCommand($connection, 'RNFR '.$path);
-        list($code) = explode(' ', end($response), 2);
+        [$code] = explode(' ', end($response), 2);
         if ((int) $code !== 350) {
             return false;
         }
 
         $response = $this->rawCommand($connection, 'RNTO '.$newpath);
-        list($code) = explode(' ', end($response), 2);
+        [$code] = explode(' ', end($response), 2);
 
         return (int) $code === 250;
     }
@@ -417,7 +418,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
         $connection = $this->getConnection();
 
         $response = $this->rawCommand($connection, 'DELE '.$path);
-        list($code) = explode(' ', end($response), 2);
+        [$code] = explode(' ', end($response), 2);
 
         return (int) $code === 250;
     }
@@ -434,7 +435,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
         $connection = $this->getConnection();
 
         $response = $this->rawCommand($connection, 'RMD '.$dirname);
-        list($code) = explode(' ', end($response), 2);
+        [$code] = explode(' ', end($response), 2);
 
         return (int) $code === 250;
     }
@@ -452,7 +453,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
         $connection = $this->getConnection();
 
         $response = $this->rawCommand($connection, 'MKD '.$dirname);
-        list($code) = explode(' ', end($response), 2);
+        [$code] = explode(' ', end($response), 2);
         if ((int) $code !== 257) {
             return false;
         }
@@ -480,7 +481,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
 
         $request = sprintf('SITE CHMOD %o %s', $mode, $path);
         $response = $this->rawCommand($connection, $request);
-        list($code) = explode(' ', end($response), 2);
+        [$code] = explode(' ', end($response), 2);
         if ((int) $code !== 200) {
             return false;
         }
@@ -497,7 +498,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
      */
     public function read($path)
     {
-        if (!$object = $this->readStream($path)) {
+        if (! $object = $this->readStream($path)) {
             return false;
         }
 
@@ -526,7 +527,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
             CURLOPT_FILE => $stream,
         ]);
 
-        if (!$result) {
+        if (! $result) {
             fclose($stream);
 
             return false;
@@ -571,7 +572,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
      */
     public function getMimetype($path)
     {
-        if (!$metadata = $this->getMetadata($path)) {
+        if (! $metadata = $this->getMetadata($path)) {
             return false;
         }
 
@@ -590,7 +591,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     public function getTimestamp($path)
     {
         $response = $this->rawCommand($this->getConnection(), 'MDTM '.$path);
-        list($code, $time) = explode(' ', end($response), 2);
+        [$code, $time] = explode(' ', end($response), 2);
         if ($code !== '213') {
             return false;
         }
@@ -601,7 +602,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
             $datetime = DateTime::createFromFormat('YmdHis', $time);
         }
 
-        if (!$datetime) {
+        if (! $datetime) {
             return false;
         }
 
@@ -650,9 +651,8 @@ class CurlFtpAdapter extends AbstractFtpAdapter
         $output = [];
 
         foreach ($listing as $item) {
-            if ($item['type'] === 'file') {
-                $output[] = $item;
-            } elseif ($item['type'] === 'dir') {
+            $output[] = $item;
+            if ($item['type'] === 'dir') {
                 $output = array_merge($output, $this->listDirectoryContentsRecursive($item['path']));
             }
         }
@@ -734,7 +734,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     protected function rawCommand($connection, $command)
     {
         $response = '';
-        $callback = function ($ch, $string) use (&$response) {
+        $callback = static function ($ch, $string) use (&$response) {
             $response .= $string;
 
             return strlen($string);
@@ -762,7 +762,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * Check the connection is established.
      */
-    protected function pingConnection()
+    protected function pingConnection(): void
     {
         // We can't use the getConnection, because it will lead to an infinite cycle
         if ($this->connection->exec() === false) {
@@ -773,14 +773,14 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * Set the connection to UTF-8 mode.
      */
-    protected function setUtf8Mode()
+    protected function setUtf8Mode(): void
     {
-        if (!$this->utf8) {
+        if (! $this->utf8) {
             return;
         }
 
         $response = $this->rawCommand($this->connection, 'OPTS UTF8 ON');
-        list($code, $message) = explode(' ', end($response), 2);
+        [$code, $message] = explode(' ', end($response), 2);
         if ($code !== '200') {
             throw new RuntimeException(
                 'Could not set UTF-8 mode for connection: '.$this->getHost().'::'.$this->getPort()
@@ -791,7 +791,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
     /**
      * Set the connection root.
      */
-    protected function setConnectionRoot()
+    protected function setConnectionRoot(): void
     {
         $root = $this->getRoot();
         if (empty($root)) {
@@ -800,7 +800,7 @@ class CurlFtpAdapter extends AbstractFtpAdapter
 
         // We can't use the getConnection, because it will lead to an infinite cycle
         $response = $this->rawCommand($this->connection, 'CWD '.$root);
-        list($code) = explode(' ', end($response), 2);
+        [$code] = explode(' ', end($response), 2);
         if ((int) $code !== 250) {
             throw new RuntimeException('Root is invalid or does not exist: '.$this->getRoot());
         }
